Updated openstack/openstack Project: openstack/cinder  1fa5a98dafa75cf1ce96e922e1d63ccf4cf54f55
Add retype in lvm driver

During retype operation, it makes sure that volume
types have same encryption, and same backend qos
before moving on.

In lvm driver, no extra_spec and qos are used. As
a result, if retype a volume to same host, it just
needs to return True, and no needs to migrate the
volume.

Change-Id: Id6273046424190ee3e36dabd4bfcf69466bf23ad
Closes-Bug: #1515840