Updated openstack/openstack Project: openstack/swift  620edc631b118317d7acf21b0ed35bf585fd1346
Fix crash when a SLO repeats a segment.

If a SLO manifest specified the same segment twice with no range on
the latter, we would crash while trying to coalesce the segment
specifications into a single request (for efficiency). Now we handle
that missing range and do the right thing.

Change-Id: If13af444ed301ebd8fd34a0d96a330ded648f0c4