Updated openstack/openstack Project: openstack/python-saharaclient  4fdb92509145eed6c7911bba95c8816ac7c112f8
Remove py26 support

As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. since we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I7920b554f541c2c40a8a3145662a0eff7faadf2f
Closes-Bug: 1519510