Updated openstack/openstack Project: openstack/python-neutronclient  4667f32103366b9c6c06c5f7b1f9fde70351177b
Remove py26 support

as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I62db55ee1a9e7564a10a7284bd0e401b2cdbc580
Closes-Bug: 1519510
Depends-On: I4a2f860e8138bddb032418ee555e07a6fa0a3f3e