Updated openstack/openstack Project: openstack/barbican  6d533310cbd9c39ce013ce0ba50a0aa46e9fe29c
Add reno for release notes management

This cycle, to address these problems for our ever-growing set of
projects, the release management team is introducing a new tool for
handling release notes as files in-tree, to allow us to simply and
continuously build the release notes for stable branch point releases
and milestones on the master branch. The idea is to use small YAML
files, usually one per note or patch, to avoid merge conflicts on
backports and then to compile those files in a deterministic way
into a more readable document for readers. Files containing release
notes can be including in patches directly, or you can create a
separate patch with release notes if you want to document a feature
than spans several patches.  The tool is called Reno, and it currently
supports ReStructuredText and Sphinx for converting note input files
to HTML for publication.  Reno is git branch-aware, so we can have
separate release notes documents for each release series published
together from the master build.

-- Excerpted from Doug Hellmann's email to the mailing list
http://lists.openstack.org/pipermail/openstack-dev/2015-November/078301.html

For more information about Reno:
http://docs.openstack.org/developer/reno/

Change-Id: I4028d52f5e67d2873b3dc9df279f78a15c35d84f