Updated openstack/openstack Project: openstack/keystonemiddleware  ecc52be445297b52d9e2f3e80a065f249e0ecb55
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I6616835bc5e13b07f79be584c4306e2573a002d8
Closes-Bug: #1489059