Updated openstack/openstack Project: openstack/neutron-lbaas  be8a085190b17458c82b8a02eb2046a2feded3d2
Setup for translation

To start translation, we need to initially import the translation
file - and place it at the proper place so that the usual CI
scripts can handle it. The proper place is for all python
projects $PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports will be done by the OpenStack Proposal bot.

Change-Id: If3a739f376b063351434bc3a5169bf9a95a5d2aa