Updated openstack/openstack Project: openstack/horizon  aa38180cf531d1c536cc8a3097eb2cb3be49bb02
Top Nav shouldn't include giant bottom margin

In the 'default' and 'material' themes, we remove the default bootstrap
margin on the top navbar because it creates a gap between the sidenav
and the topnav. This should be a global style so that other themes can
take advantage of it. It looks best with a single 1px value on
alternate themes.

Change-Id: Ie0f3e5e22c29fe737d128af6d2fd6499fe9ee99a
Closes-Bug: #1518455