Updated openstack/openstack Project: openstack-infra/project-config  679b292d27f5ff71c4ed599ac519e430910cd8b2
Pull latest changes in run-layout.sh

Some of my local tox tests were failing in odd ways; eventually I
chased it down to out-of-date versions of tools left in ".test"
directory.

This tries to pull the latest changes if the repos exist, and adds a
few helpful output points to let you know what's going on.

Change-Id: I30b7ddf84520e8895466a49ed6481812ea8875a8