Updated openstack/openstack Project: openstack/manila  3b1e0b34d55587cc16a5a001344c55e8e4b05c58
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: Id54bb80944385579cfa136fce78cb5be55393f81
Closes-Bug: #1489059