Updated openstack/openstack Project: openstack/python-heatclient  ed250fbcfea22d6180cecca7f5f02d566d5fc7a5
Remove py26 support

As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.
Closes-Bug: 1519510

Change-Id: I680043c48c75a50b3ef8eed4f22e97dca029c477