Updated openstack/openstack Project: openstack/nova-specs  4194d96c38612392962ccc0589df6f346ad6ac92
Allow volume operations when shelved

Volume operations are allowed when an instance is
shutdown, paused and even soft deleted, but not when
shelved or shelved_offloaded. This spec will enable them
when shelved and shelved_offloaded as well.

Besides being useful in itself, this is also a first
step towards allowing operations on the boot volume
when shelved.

blueprint volume-ops-when-shelved
APIImpact
Change-Id: I52f32831bad5f5e1ff378c51186ee798dc3efa2e