Updated openstack/openstack Project: openstack/nova-specs  7545f0e939ffe5d1fec34457823bf4101d02fb47
Detach and attach boot volumes

It is sometimes useful for a cloud user to be able to
detach and attach the boot volume of an instance that
is not running. Currently nova does not allow this and
some operations assume it does not happen. This spec
proposes allowing the detach and attach of boot volumes
in the shelved_offloaded state and adding safeguards to
ensure it is safe.

blueprint detach-boot-volume

APIImpact

Change-Id: I8a0de69700e97e955a2358e4c97bd1d9b6e71117