Updated openstack/openstack Project: openstack/glance_store  5c15a32a4eb51ff64b023520a97740dc0a7e79d3
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I2bd65c7af7c87b6c36439e6e4bb6230eca8d33ee
Closes-Bug: #1489059