Updated openstack/openstack Project: openstack/python-novaclient  f02ca48a4efdd7c6969f0a9efb0ed6679b289213
Fix H404/405 violations in client.py,base.py,api_version.py

There is a lot of H404/405 violations in novaclient, and it is better
to fix those to have a better doc string for class/methods.

This patch fixes these violations for below files-
-api_versions.py
-base.py
-client.py

As there are lot of violations and cannot be fixed in single patches,
So separating those in multiple patches for easy review.

Change-Id: I30a714fc3b0b317f7ffa4a99fbb224b4d5f5477b
Partial-Bug: #1521899