Updated openstack/openstack Project: openstack/oslo.config  5490a28a1398f10a3d346fef23374a972f90c40f
[fix-compat] Support older use case for Opt type

In I7cf94a1b2cbbfe67282e8a157548caff825d2811, we we moved
type formatting from generator to type, however we still have
some older code/libraries that were using a Opt with a type
set to a string. We need to support them until we find and
fix all use cases. So we explicitly check if format_defaults
is present and call it if it is. If format_defaults is
not present we call the older code from the change id
mentioned above.

Closes-Bug: #1522637
Change-Id: I298e75afbd43bd35aba93c38129008789ed39146