Updated openstack/openstack Project: openstack/keystone  45c19fcd8c4cc382a7471432cd9f72b809e1d5b1
Simplify LimitTests

Limit tests currently used for SQL backend testing are too complex.
Due to this complexity they could be used for SQL only. They assumed
possibility to create a domain, while it is not possible when LDAP
identity backend is used with multidomain support.

Simplify LimitTests and assume that a default domain is used for all
operations.

Change-Id: I77d581e03dae86c299ce2e1e56164dde3c4ad72a