Updated openstack/openstack Project: openstack/python-ceilometerclient  f247dde3570eb3d314df80e169232a716e6679e1
Remove py26 support

as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I2218cf3464a9c1667f8eb812a519b2bc3ffe4d0d
Closes-Bug: 1519510