Updated openstack/openstack Project: openstack-infra/puppet-gerrit  4580769b43fd62726d28ac059f4db184f842cac8
Ensure all groups are visible to all registered users.

If a group is not visible to the currently logged in user then
Gerrit will deny that user the ability to query for changes by
that group[1].  For example some users may not be able to make
the type of query specified in change Ief5408ee4774f94

Set newGroupsVisibleToAll[2] to true by default so that all
registered users can query by group.

[1] https://review-dev.openstack.org/Documentation/intro-project-owner.html#groups
[2] https://review-dev.openstack.org/Documentation/config-gerrit.html#groups

Change-Id: Iab5d72874ff620c1aca7041291bc765854fec87b