Updated openstack/openstack Project: openstack/neutron-lbaas  1e155cc593f82866c91e10306b8bdf950a552252
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I7e279cbb3a6c4710b8361c8cc2df368bd28c3203
Closes-Bug: #1489059