Updated openstack/openstack Project: openstack/python-zaqarclient  c792f9b7d962290c98f37f63ed225c4129b4ebf1
Remove py26 support

As of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: I6420a95a58c85830412ae5792208d952116ca7e6
Closes-Bug: 1519510