Updated openstack/openstack Project: openstack/glance  b9f9000bc31becf03484ccc8f650baa83adf77aa
Change the default notification exchange to glance

As we agreed that we'll change the default notification exchange name
from 'openstack' to 'glance' in M release.

The openstack convention is to send out all the notifications to the
exchange with the same name as the project name, i.e. nova sends all its
notifications to the exchange 'nova', cinder sends all its notifications
to the exchange 'cinder', etc. In such way, other projects interested in
those notifications (e.g. ceilometer) can listen on different exchanges
for different notifications.

Change-Id: I658fb0896aecea4fd5db1a9d3a8bbb64a91a7224
Closes-Bug: #1467776