Updated openstack/openstack Project: openstack/nova-specs  e3e575323217be042d4ab430a300642b88b612a9
Add a part of api-wg guideline

During nova-specs reviews, it is difficult to check all API changes
for following api-wg guideline because there is a lot of nova-specs
patches. There are some common mistakes against following guideline[1],
then this patch adds a part of the guideline to the template file for
avoiding these mistakes.

  REST API resource names
    * Resource names may include hyphens.
    * Resource names should not include underscores or other punctuation
      (sole exception is the hyphen).
  Fields in an API request or response body
    * Field names should use snake_case style, not CamelCase or MixedCase
      style.

[1]: https://github.com/openstack/api-wg/blob/master/guidelines/naming.rst

Change-Id: I20039ad1f7882cd5341431690b82dadab3cd1588