Updated openstack/openstack Project: openstack/debtcollector  0132731a3819256902ecd350da72233bb4b00b98
Add 'moved_read_only_property' descriptor

For when a existing property (class property or
instance property) has been renamed it is quite
useful to be able to deprecate the old name/location
and let users that access that property know when
they read the property what the new location is.

This adds a new descriptor that emits such a
warning in a format that matches the rest of the
debtcollector functions/decorators/descriptors.

Change-Id: I9305ba512856e957c06996b637acb4a8c4fa5d04