Updated openstack/openstack Project: openstack/python-keystoneclient  9f9fd47919f91a3523bc1904f69179b7eb22ad25
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I0cc290e9fc13c8d61d933fe4b457008df0a29c5a
Closes-Bug: #1489059