Updated openstack/openstack Project: openstack-infra/shade  e9cd56cefca92dd063f2658132d1da79b51df736
Change the client imports to stop shadowing

Because sometimes the easter bunny wants to kill you, we have been
importing the client library contructor modules with a name that is
identical to the proprety we set on the class to provide the client
object. That is bad practice in any circumstance.

Since we only use those things in one and only one place, don't bother
coming up with a better alias for them - just import the module path
itself directly. Explcit good.

Change-Id: Ib0d7e616df1304277b4b9936a64793bcba8787e3