Updated openstack/openstack Project: openstack/neutron  5fdab56c36644150f9b6de5235e26d04d24ece46
Add instrumentation devref, Part I

Presents what instrumentation is available from VIFs in Nova,
Metering Lables and Rules, Linux Bridge, and OVS. Proposes
mappings for structures defined in RFC 2863 and RFC 4293 and
the method that will be followed for a data collection proof
of concept.

How to aggregate and consume these counters will be covered
in future patch sets that extend this devref.

Change-Id: I6c1ad0c4cf60d0069c5e057d77f75c12b04a020c
Partial-bug: #1475736