Updated openstack/openstack Project: openstack/keystone  d7910d685f0095cc203e71b0afa9f6d0b2a2af08
Put py34 first in the env order of tox

To solve the problem of "db type could not be determined" on py34 we
have to run first the py34 env to, then, run py27. This patch puts py34
first on the tox.ini list of envs to avoid this problem to happen.

Change-Id: I36ec1829229081b6bc935ecc0d27836c8699a753
Closes-Bug: #1489059