Updated openstack/openstack Project: openstack-infra/project-config  06693519c56b9bcbbf5b9306ee0af4af7ff9352a
remove gate-devstack-dsvm-cells

gate-devstack-dsvm-cells was a stop gap job on juno/kilo to use
devstack exercises on cells until the tempest job got working. The
tempest job was working in kilo, and now votes.

With juno gone, and the old job almost testing nothing on kilo (it
boots 1 vm only), we should remove it and just leave the tempest-cells
job where it already is.

Change-Id: I829903b0690faea2d499f4ba8a2393cf31ed4601