Updated openstack/openstack Project: openstack/python-neutronclient  04aa749b828c1c1055fefb684e9a71a5db04eb57
Add release notes for changes since 3.0.0 release

We decided to use reno for coming releases, but we don't
have entries for changes made after the previous release
was shipped. This commit adds notes for them.

Also fixes the wrong place of "Release Notes" section in
the document index. It was my mistake in the past commit.

Change-Id: Ie0eedf8cb82a57db070a65f73d965f1a3ce378ab