Updated openstack/openstack Project: openstack/neutron  47dce8f40575622602620e663df5b5155c1b8cdb
Hyper-V: remove driver from the neutron tree

The hyperv drivers and code should be part of the networking-hyperv
project (https://github.com/openstack/networking-hyperv).

A few changes are necessary in order to prevent Hyper-V deployments from
breaking, especially when upgrading to Mitaka.

Hyper-V deployments are configured to use the in-branch HyperVSecurityGroupsDriver.
Removing it will cause the Hyper-V Neutron Agent to fail. If the agent is
configured to use the old driver, the networking_hyperv's driver must be used
instead and the users must be warned to update their configuration files to
use the networking_hyperv's security groups driver.

Removes the neutron-hyperv-agent entry point from setup.cfg.
Removes the hyperv mechanism_driver from setup.cfg
Moves the in-tree HyperVSecurityGroupsDriver to the networking_hyperv equivalent.
Logs a warning if the in-tree HyperVSecurityGroupsDriver is used.
Removes pywin32 and wmi requirements, as they've been included in networking_hyperv
and they are Hyper-V specific requirements.
Adds release note regarding the deprecated security groups driver.

Co-Authored-By: Claudiu Belu <cbelu@cloudbasesolutions.com>

Depends-On: I3a25f18b4f3a0621cb92b44eb61e434fa87e0aab

Change-Id: I32451cba6933e88306a4308a07f3f0d1b81f626c
Closes-bug: #1520054