Updated openstack/openstack Project: openstack/neutron-fwaas  b328b91c50039a57bca2c84183d0223f379af14d
Remove dependency on neutron for topics

FWaaS uses L3_AGENT and FIREWALL_PLUGIN topics. The latter is only
used in FWaaS, and can be removed from neutron later.

The goal here is to chip away at removing dependencies on the neutron
repo. This is just one simple case to clean up some unneeded references.

Placed the literals in a new file, which can be used for storing other
constants, as desired.

Change-Id: Idcfeee6ddc2982fff49f44064c918bfd4627be12
Implements: blueprint neutron-lib