Updated openstack/openstack Project: openstack/neutron  13a81ab71d7c98ab3d9e51c8bc253a2bc0c87a8c
Delete stale neutron-server manual

The configuration options come from oslo and the server
executable is usually wrapped in a service script, supplied
by packagers and/or deployment tools. Any extra documentation
available in tree is of relative value, and the fact that
this file has been virtually ignored ever since it was
added is a testament of that.

Let's stop its agony and wish it to rest in peace.

Closes-bug: #1520041

Change-Id: If5bba557526903b8064ee28628a21c3459ca85bc