Updated openstack/openstack Project: openstack/python-openstackclient  42c5d2ee1b40b9eb82a44ac938fbe108775e0250
Refactor network test: Introduce TestNetworkv2 and TestNetwork to improve unit test of network

The class inherit architecture in network unit test is different from other test classes,
which leads to lots of redundant code. This patch will make it the same as the other test
classes. And it will be more convenience for the coming up refactor.

Change-Id: I6f239dd54b9401ff2bbcf7ffdeb18769a450f573
Implements: blueprint osc-network-unit-test-refactor
Related-to: blueprint neutron-client