Updated openstack/openstack Project: openstack/horizon  1b6807baf385d6e5768f149fa8e4d07bc24ebff1
Exclude network which don't have subnet

In create network form, the given network should
have a subnet to create an instance. otherwise the
create call will fail.

We can filter the network list based on the subnet
presence in the network to make the launch instance
successful and avoid unnecessary server call.

Change-Id: Ia37f32d5972ab84530a516dd488014081861c630
Closes-Bug: #1479722