Updated openstack/openstack Project: openstack/api-site  9ea27858619320e6fd8253d4239bb8550ced798e
Fix the link of "Configures extra routes"

The links of "Configures extra routes" and "Update router" are
duplicated, and we cannot open the link of "Configures extra routes"
on api-site, because "Update router" is always open due to the order.
This patch fixes this problem.

In addition, this patch changes the title("Update router") because
that is not right on the API behavior.

Change-Id: Ib8c965cef7d9e49ef7d03cbb0234f15826345cf5
Closes-Bug: #1522691