Updated openstack/openstack Project: openstack/ceilometer  ce6c29a0de726e4389cfd3e325094c3c63c8e5ec
MAINTAINERS: remove outdated data

This is outdated and never updated anyway. It does not reflect the reality
anymore, so let's stop trying to have that as it does not give any useful
information anyway. We're a team, we maintain everything!

(and we can use git log to find out things)

Change-Id: I28dd5fdfdc0ef4b0d9816b38572c5bbac3d6f9cf