Updated openstack/openstack Project: openstack/neutron  798c8d780f8c9f5c45484a6d523f33702f1f6ec8
Tox: Remove fullstack env, keep only dsvm-fullstack

In the functional tests we have 'functional' and 'dsvm-functional'
venvs. The difference is that dsvm-functional requires can run
rootwrap commands on the machine. In the fullstack context,
not running rootwrap commands doesn't make sense, as fullstack
tests are explicitly and solely integration tests.

Change-Id: I1e85537d137a1b773198facbb835a65b6d0ce9b4