Updated openstack/openstack Project: openstack/python-openstackclient  695500496cb0b5682d8de8d60fc74ffe311445f8
Use formatter in server.py for "server list" tests.

We should use the formatter in the original code to set the
expected data in each test case.

The same problem exists in other test cases, like "server create".
But data structure in TestServerCreate is different from TestServerList,
so will fix the problem in another patch.

Change-Id: I233f7a91cd4cc1e996941e26ea85490fa3290572