Updated openstack/openstack Project: openstack-dev/pbr  379e42abde70300e426fb4e8d5115c67fd2e45b9
Add patch to properly get all commands from dist

Distribution().get_command_list() is not hooked by setuptools like
print_commands() is, as such we do not get a complete command list.

Here we perform the hook ourselves to prevent the need to depend on
a super new version of setuptools.

Change-Id: Ie6088aedb92713be4b8b9af4a4ce3d300a114d7b