Updated openstack/openstack Project: openstack/heat  b611e911c68abb9c8f9ac34b1e17147b7cfaf02a
Fix to send only one notification when update completes

Presently when an update is COMPLETE/FAILED, two events are reported.

In update_task we are already sending notification when an update
completes, however due to recent changes, notifications are also
sent at the time of releasing stack lock.

For update, we need not send notification when releasing lock.

Change-Id: Idc81f87b143dc43fa5409798534646d8400559ff
Closes-Bug: #1516090