Updated openstack/openstack Project: openstack/python-cinderclient  7970e0fb7ebd70b65238d22968a9f22b87772ed0
Remove py26 support

as of mitaka, the infra team won't have the resources available to
reasonably test py26, also the oslo team is dropping py26 support
from their libraries. sine we rely on oslo for a lot of our work,
and depend on infra for our CI, we should drop py26 support too.

Change-Id: Icf1b13b044de03705688df929ed74429879c6dd8
Closes-Bug: 1519510