Updated openstack/openstack Project: openstack/ironic  e1417ff476267f4854d6e5b1d476d416b154cc21
Split BaseConductorManager from ConductorManager

This is the second patch in a chain of refactoring patches in preparation
for moving to Futurist library. See blueprint futurist for details.

This patch moves common conductor initialization and lifetime code,
and some common conductor methods to a separate class BaseConductorManager.
I expect most of the work in blueprint futurist to touch this class.

Implements: blueprint futurist
Change-Id: If4f018d6e0473e62134a0427a0ad0ea733c10ea4