Updated openstack/openstack Project: openstack-infra/puppet-subunit2sql  5327cc64a87f63958fd24929bb5d2765c61f752e
Handle 404 and download errors properly

This commit adds proper error handling to the subunit gearman worker
in the event a download error occurs. Prior to this change if a
subunit stream is unable to be downloaded for whatever reason an empty
row was being added to the subunit2sql db. This was because after
logging the exception an empty stream file gets passed to subunit2sql
which would treat that as adding a run which didn't run anything.
This a waste of time and is confusing to users of the DB because
there will be a number of runs in the DB which didn't actually run
anything.

Change-Id: I1f8fd7ffd9c16ce2dddd534d4c641e6d65249d91