Updated openstack/openstack Project: openstack-infra/project-config  df3fdef15de58574ef4452c72e313fe8b0730e37
Move keystonemiddleware to use the same ACLs as keystoneauth

After discussing this with steve martinelli (PTL), it makes sense to
move the keystonemiddleware core ACLs to have the same group that
really focuses on keystoneauth. This is because both keystoneauth
and keystonemiddleware are mostly looked at by a slightly more
specialized group.

Keystone-core group still has full rights to vote on the project. This
allows for an expanded set of eyes who focus more on authz within
openstack to be involved with keystonemiddleware.

Change-Id: If56f5b7636a6880f61ac9510f2bff546b7c2bf4b