Updated openstack/openstack Project: openstack/requirements  2f01f1d3822caee757f8351d35e6a9377efb4dd6
Add funcsigs into global-requirements.txt

This library will be useful for oslo.utils and its consumers
so that they can correctly examine function signatures across
the python versions that the community supports.

Is the library actively maintained?

  Not especially, lifeless and I are working on taking over
  ownership, this should be happening very soon as it appears
  the author is not responsive to emails or other.

Is the library good code?

  I have seen worse.

Is the library python 3 compatible?

  The main purpose of this library is to be a backport of
  PEP 362 so on python 3 it isn't really needed in the first
  place.

Is the library license compatible?

  Yes.

Is the library already packaged in the distros we target?

  It appears to be packaged in ubuntu latest, fedora appears
  to have a package that may not be distributed yet as well.

Is the function of this library already covered by other
libraries in global-requirements.txt?

  Nope

Change-Id: I1aa0054089ca57fc2d68779f4ee133a9750fec2a