Updated openstack/openstack Project: openstack/openstack-manuals  13e2714c3caff406de3b99592a1cdbaef299aec7
[install] Fix heat trustee configuration

The current example config is broken with the default deferred
authentication of trusts - the project_domain_id and project_name
cannot be specified in this section or keystone will throw a
"cannot be scoped to multiple targets" error when we attempt to get
a token scoped to a trust.

Also, the auth_uri is unused by the keystoneclient password auth
plugins, so remove it (it has been a source of confusion in the
referenced bug).

Change-Id: I6305be3c693993de4d16d82fd6de936e92c437c5
Closes-Bug: #1300246
Backport: liberty