Updated openstack/openstack Project: openstack/keystone-specs  2509e5b2406a08ef1d63dfd0f5b85623cb2a4d32
Expand endpoint filters to service providers

Endpoint filtering is a keystone extension since the Havana release,
in Juno it was expanded by the Multi-Attribute Endpoint Grouping
spec that improves the filtering flexibility.

Since Kilo, we added the concept of Service Providers to keystone,
the list of such resources is returned in the token alongside the
service catalog.
In Mitaka we are proposing to extend endpoints filtering to also
support Service Providers. This should not bring API nor model
changes, just extra checks to the endpoint filter controller/backend.

bp service-provider-filters

Co-Authored-By: Marek Denis <marek.denis@cern.ch>
Co-Authored-By: Iury Gregory Melo Ferreira <iurygregory@gmail.com>

Change-Id: Ice20ede8a27eefad1a98d810dfee425a71d06a86