Updated openstack/openstack Project: openstack/keystone  3843a8f1c6b80eb89dddd9931bca5d908fd4f066
Deprecate ldap Role

The LDAP backend for assignment and resource drivers has been
deprecated since the 'K' release. It's been decided that its time
to deprecate LDAP role, since without the support for assignment
and resource, the role backend is likely broken.

The LDAP role backend was split from assignment backend and should
have also been deprecated according to the agreement from the keystone
team, but it looks like it was missed [1].

[1] http://lists.openstack.org/pipermail/openstack/2015-January/011337.html

Change-Id: I390f5c2fb36d617a83fce618708febca3da9f1d2
implements: bp deprecated-as-of-mitaka