Updated openstack/openstack Project: openstack/requirements  0a9dfce58e041438a741ab66359e95baffe28ff9
Add notifier taskflow extraction library

This library/chunk of code is extracted out of
taskflow and moved to be its own independent
library (it shares the same license as taskflow
and will likely soon be an oslo library).

Code is at (for now):

https://github.com/harlowja/notifier

This is needed for review:

https://review.openstack.org/#/c/228892/

Docs are at (for now):

http://notifier.readthedocs.org/

----

Is the library actively maintained?

Yes.

Is the library good code?

Obviously yes.

Is the library python 3 compatible?

Yes.

Is the library license compatible?

Yes.

Is the library already packaged in the distros we target?

No (not directly, they package taskflow, where
this code came from).

Is the function of this library already covered by
other libraries in global-requirements.txt?

No.

Change-Id: If63e54b4eee77361a4c5c53b724db0f600fddd37