Updated openstack/openstack Project: openstack/requirements  5e4032a4ca7e71b74825db596e9603a6bb899982
Update to pylint 1.4.5.

Pylint 1.4.5 is a dependency-only-change - it caps astroid, which
pylint uses to parse Python code. astroid 1.4.0 was backwards
incompatible. Switch to using 1.4.5, so that we can eventually unpin
astroid in preparation for projects moving to pylint 1.5 in future.

We cannot remove the astroid cap until all projects are using 1.4.5,
otherwise projects that are still dealing with the astroid mess will
be unable to sync a cap for it. Once all projects are on 1.4.5, we can
remove it trivially.

Change-Id: I1ff29325947723c7136272bc873cafa769619063