Updated openstack/openstack Project: openstack-infra/storyboard  708f6f2177f84baa5bc70bf433a9ab21a8743bf7
Don't delete subscription if sub_resource is set

This patch stops subscriptions being erroneously deleted if a DELETE
request is recieved with a path which has a sub_resource but no
sub_resource_id. For example:

    DELETE /v1/stories/1/tags

This fixes a bug whereby deleting a tag from a story unsubscribes
everyone from that story.

Change-Id: I2166ae63d940ec40e45a8bb7a97779099131f7cd