Updated openstack/openstack Project: openstack-infra/project-config  c7f5f1e0301b9b44e82a926e40a1e364b5b5cefa
Run ceilometer-python34 on all branches

The job was marked to run only on master in
Ifef79456e7caea8ecc5dda6b0ea8b6869eb99f93 - at that time so that it did
not run on juno. Since juno is eol'ed, let's run the job
on all branches now and thus remove the filter.

Change-Id: Ib38ce0a8baf7458f897a44eca4efc280a27a5444